Rename http-service.rb to http_service.rb
