+++ lib/http_service.rb
